Structure changes to further support easy editing for JSR implementation
