+++ simulations.h
@@ -1,9 +1,5 @@
-//
-// Created by jayle on 3/15/2024.
-//
-
-#ifndef CPU_SIM_MAIN_H
-#define CPU_SIM_MAIN_H
+#ifndef CPU_SIM_SIMULATIONS_H
+#define CPU_SIM_SIMULATIONS_H
 
 typedef unsigned char Byte;
 typedef unsigned short Word;
@@ -17,6 +13,14 @@ typedef unsigned int u32;
 #define LDY_im  ((Byte) 0xA0)  //Load Y Register (Immediate Mode)
 #define JSR     ((Byte) 0x20)  //Jump to sub-Routine
 
-u32 numCycles = 0;
+u32 numCycles = 3;
 
-#endif //CPU_SIM_MAIN_H
+void reset(struct CPU *pCpu, struct Memory *pMemory);
+
+Byte fetchInstruction(Word *pCounter, struct Memory *pMemory, u32 *pCycles);
+
+Byte readInstruction(Byte address, struct Memory *pMemory, u32 *pCycles);
+
+void exec(struct CPU *pCpu, struct Memory *pMemory, u32 *pCycles);
+
+#endif //CPU_SIM_SIMULATIONS_H
